1. Simple cleanup and refactor 2. Add doc strings
