+++ src/components/chartComponents/index.jsx
@@ -1,9 +1,10 @@
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faXmark, faRotateLeft, faCircleInfo, faExternalLink } from '@fortawesome/free-solid-svg-icons';
-import { useState } from 'react';
+import { useState, useEffect } from 'react';
 import { useChart } from '../../context/chartContext';
 
 export const ChartInstruction = () => {
+  // Displays instruction for interacting with the chart when the info icon is hovered
     const [showInstructions, setShowInstructions] = useState(false);
     return (
       <div id="chart-instructions-container">
@@ -21,13 +22,18 @@ export const ChartInstruction = () => {
     );
   };
   
-const ZoomResetTool = ({ chart }) => {
-    const handleResetZoom = () => {
+const ZoomResetTool = () => {
+  // Resets the zoom level of the chart when the reset zoom icon is clicked
+  // Uses the useChart hook to access the chart object
+  const { chart } = useChart();
+
+  const handleResetZoom = () => {
     if (chart.current) {
       chart.current.resetZoom();
     }
   };
-    return (
+
+  return (
     <FontAwesomeIcon
       id="zoom-reset-button"
       icon={faRotateLeft}
@@ -38,6 +44,8 @@ const ZoomResetTool = ({ chart }) => {
 };
 
 const CloseButton = ({ handleClose }) => (
+  // Closes the chart when the close icon is clicked
+  // Receives the handleClose function as a prop
   <FontAwesomeIcon
     id="chart-close-button"
     icon={faXmark}
@@ -46,22 +54,31 @@ const CloseButton = ({ handleClose }) => (
   />
 );
 
-const DataAccessTool = ({ dataAccessLink }) => (
-  dataAccessLink && (
-    <a id="data-access-link" href={dataAccessLink} target="_blank" rel="noreferrer">
+const DataAccessTool = ({ dataAccessLink }) => {
+  // Displays a link to the data access page when the data access link is provided
+  const [link, setLink] = useState(dataAccessLink);
+
+  useEffect(() => {
+    setLink(dataAccessLink);
+  }, [dataAccessLink]);
+  
+  if (!link) return null;
+
+  return (
+    <a id="data-access-link" href={link} target="_blank" rel="noreferrer">
       <FontAwesomeIcon icon={faExternalLink} />
     </a>
-  )
-);
+  );
+};
   
 export const ChartTools = ({ dataAccessLink, handleClose }) => {
-  const { chart } = useChart();
-
+  // Displays the chart tools
+  // Receives the data access link and handleClose function as props
   return (
     <div id="chart-tools-right">
       <div id="chart-controls">
         <DataAccessTool dataAccessLink={dataAccessLink} />
-        <ZoomResetTool chart={chart} />
+        <ZoomResetTool />
         <CloseButton handleClose={handleClose} />
       </div>
     </div>
