+++ src/pages/dashboard/index.jsx
@@ -106,5 +106,3 @@ export function Dashboard({
   );
 }
 
-
-
